Forward requests asynchronously
The ecowitt gateways aren't powerful devices so return a response right away and forward to the other targets out of band.

This way if the ecowitt is configured to send every 16 seconds but it takes up to 5 seconds to hit many targets, it won't take longer than the 16 second interval which ecowitt will poll the endpoint.